refactor: get-most-followed-mbmer
